feat(visualizer): add 2D plotting class
Implement Plot2D as the plotting class. It may be utilized
by each trainer iteration to store and update plot image in
the active checkpoint directory, per metric. One object may be
defined per one metric. Supports the concatenation of additional
measurements. Initialized/added measurements must be in a
format similar to training-history.json's metrics_measurements,
however only for one metric, and with the indices of dataset and
metric inverted. For the latter, implement a utility method of
CheckpointStorageManager called invert_dataset_metric to provide the
inversion utility.